+++ src/main/groovy/com/trevorism/data/service/lookup/DatastoreLookupService.groovy
@@ -2,6 +2,7 @@ package com.trevorism.data.service.lookup
 
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
+import com.trevorism.data.DataUtils
 import com.trevorism.data.model.SingleDatasourceRequest
 import com.trevorism.data.model.exception.InvalidLookupException
 import com.trevorism.https.SecureHttpClient
@@ -10,8 +11,9 @@ class DatastoreLookupService implements LookupService {
 
     private SecureHttpClient client
     private Gson gson = (new GsonBuilder()).setDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'").create()
-    private final static String DATASTORE_BASE_URL = "https://datastore.data.trevorism.com/object"
-    private final static String BIGQUERY_BASE_URL = "https://bigquery.data.trevorism.com/object"
+    private static final String DATASTORE_OBJECT_BASE_URL = "${DataUtils.DATASTORE_BASE_URL}/object"
+    private static final String BIGQUERY_OBJECT_BASE_URL = "${DataUtils.BIGQUERY_BASE_URL}/object"
+    private static final String MEMORY_OBJECT_BASE_URL = "${DataUtils.MEMORY_BASE_URL}/object"
 
     DatastoreLookupService(SecureHttpClient client) {
         this.client = client
@@ -27,9 +29,11 @@ class DatastoreLookupService implements LookupService {
         String datasource = parts[0]
         String kind = parts[1]
 
-        String url = "$DATASTORE_BASE_URL/$kind"
+        String url = "$DATASTORE_OBJECT_BASE_URL/$kind"
         if (datasource == "bigquery")
-            url = "$BIGQUERY_BASE_URL/$kind"
+            url = "$BIGQUERY_OBJECT_BASE_URL/$kind"
+        if (datasource == "memory")
+            url = "$MEMORY_OBJECT_BASE_URL/$kind"
 
         String json = client.get(url)
         return gson.fromJson(json, List)